+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Remove default text `-` when reactivating NumberOption.
-- Focus NumberOption input when after activating it.
+- Focus NumberOption input after activating it.
 
 ## [0.7.1] - 2020-06-02
 ### Added