+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `LocationSearch` component now send queries to `vtex.geolocation-graphql-interface` instead of `places-graphql`.
 
 ## [0.10.0] - 2020-09-01
 ### Added
-- `LocationSearch`
+- `LocationSearch`.
 
 ## [0.9.1] - 2020-08-18