+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.13.1] - 2020-11-19
 ### Fixed
 - `LocationCountry` label is now displayed.
 - Failure message will no longer show up when starting to type on `LocationSearch`.